+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2022/07/29
+### Changed
+- Minimum image width and height are back to 4 pixels.
+- Retain original image dimensions when converting from PNG.
+### Fixed
+- Fix issue with converting images that are not a power of 2.
+
 ## [0.2.1] - 2022/07/20
 ### Added
 - Add `DdsConversionOptions` to make mipmaps and shuffling configurable.